Changed assertion system out for exceptions. Removed lots of `noexcept` specifiers to make up for the new exception system. Fixed a bug in Allocator interface regarding integer-type. Renamed ParsedFileInfo to just FileInfo Renamed GLFormats to GLTools and moved it into detail headers. Renamed functions including "Mipmap" or "MipLevel" to use just "Mip" Renamed VkFormats to VulkanTools and moved it into detail headers. Started transitioning to east-const codestyle. Changed some integer-type to std::uint_least32_t in the backend

+++ include/Texas/Allocator.hpp
@@ -13,7 +13,7 @@ namespace Texas
 	class Allocator
 	{
 	public:
-		virtual std::byte* allocate(std::uint64_t amount) = 0;
+		virtual std::byte* allocate(std::size_t amount) = 0;
 		virtual void deallocate(std::byte* ptr) = 0;
 	};
 }
